fix: linting errors (ineffassign, imports, dupl)
