+++ walletapi/balance_decoder.go
@@ -16,17 +16,16 @@
 
 package walletapi
 
-import "runtime"
-import "fmt"
-import "sort"
-import "math/big"
-import "encoding/binary"
+import (
+	"encoding/binary"
+	"fmt"
+	"math/big"
+	"runtime"
+	"sort"
 
-//import "github.com/mattn/go-isatty"
-//import "github.com/cheggaaa/pb/v3"
-
-import "github.com/deroproject/derohe/cryptography/crypto"
-import "github.com/deroproject/derohe/cryptography/bn256"
+	"github.com/deroproject/derohe/cryptography/bn256"
+	"github.com/deroproject/derohe/cryptography/crypto"
+)
 
 // this file implements balance decoder whih has to be bruteforced
 // balance is a 64 bit field and total effort is 2^64
