+++ walletapi/daemon_communication.go
@@ -22,39 +22,27 @@ package walletapi
  *
  * *
  */
-//import "io"
-//import "os"
-import "fmt"
-import "time"
-import "sync"
-import "bytes"
-import "math/big"
+import (
+	"bytes"
+	"context"
+	"encoding/hex"
+	"fmt"
+	"math/big"
+	"runtime/debug"
+	"strings"
+	"sync"
+	"time"
 
-//import "bufio"
-import "strings"
-import "context"
-
-//import "runtime"
-//import "compress/gzip"
-import "encoding/hex"
-
-import "runtime/debug"
-
-//import "github.com/vmihailenco/msgpack"
-
-//import "github.com/gorilla/websocket"
-//import "github.com/mafredri/cdp/rpcc"
-
-import "github.com/deroproject/derohe/rpc"
-import "github.com/deroproject/derohe/block"
-import "github.com/deroproject/derohe/config"
-import "github.com/deroproject/derohe/globals"
-import "github.com/deroproject/derohe/cryptography/crypto"
-import "github.com/deroproject/derohe/errormsg"
-import "github.com/deroproject/derohe/transaction"
-import "github.com/deroproject/derohe/cryptography/bn256"
-
-import "github.com/creachadair/jrpc2"
+	"github.com/creachadair/jrpc2"
+	"github.com/deroproject/derohe/block"
+	"github.com/deroproject/derohe/config"
+	"github.com/deroproject/derohe/cryptography/bn256"
+	"github.com/deroproject/derohe/cryptography/crypto"
+	"github.com/deroproject/derohe/errormsg"
+	"github.com/deroproject/derohe/globals"
+	"github.com/deroproject/derohe/rpc"
+	"github.com/deroproject/derohe/transaction"
+)
 
 // this global variable should be within wallet structure
 var Connected bool = false
@@ -705,7 +693,7 @@ func (w *Wallet_Memory) synchistory_block(scid crypto.Hash, topo int64) (err err
 
 	EWData := fmt.Sprintf("%x", current_balance_e.Serialize())
 
-	previous_balance = w.DecodeEncryptedBalance_Memory(previous_balance_e, 0)
+	_ = w.DecodeEncryptedBalance_Memory(previous_balance_e, 0)
 	current_balance = w.DecodeEncryptedBalance_Memory(current_balance_e, 0)
 
 	// we can skip some check if both balances are equal ( means we are ring members in this block)
