+++ walletapi/daemon_connectivity_default.go
@@ -1,0 +1,17 @@
+package walletapi
+
+import (
+	"net/http"
+
+	"github.com/creachadair/jrpc2"
+	"github.com/gorilla/websocket"
+)
+
+var netClient *http.Client
+
+type Client struct {
+	WS  *websocket.Conn
+	RPC *jrpc2.Client
+}
+
+var rpc_client = &Client{}
