+++ walletapi/wallet_transfer.go
@@ -16,33 +16,16 @@
 
 package walletapi
 
-import "fmt"
+import (
+	"encoding/hex"
+	"fmt"
 
-//import "sort"
-//import "math/rand"
-//import cryptorand "crypto/rand"
-
-//import "encoding/binary"
-import "encoding/hex"
-
-//import "encoding/json"
-
-//import "github.com/vmihailenco/msgpack"
-
-import "github.com/deroproject/derohe/config"
-import "github.com/deroproject/derohe/cryptography/crypto"
-
-//import "github.com/deroproject/derohe/crypto/ringct"
-import "github.com/deroproject/derohe/transaction"
-
-//import "github.com/deroproject/derohe/globals"
-import "github.com/deroproject/derohe/rpc"
-
-//import "github.com/deroproject/derohe/ddn"
-
-//import "github.com/deroproject/derohe/structures"
-//import "github.com/deroproject/derohe/blockchain/inputmaturity"
-import "github.com/deroproject/derohe/cryptography/bn256"
+	"github.com/deroproject/derohe/config"
+	"github.com/deroproject/derohe/cryptography/bn256"
+	"github.com/deroproject/derohe/cryptography/crypto"
+	"github.com/deroproject/derohe/rpc"
+	"github.com/deroproject/derohe/transaction"
+)
 
 /*
 func (w *Wallet_Memory) Transfer_Simplified(addr string, value uint64, data []byte, scdata rpc.Arguments) (tx *transaction.Transaction, err error) {
@@ -224,7 +207,7 @@ func (w *Wallet_Memory) TransferPayload0(transfers []rpc.Transfer, ringsize uint
 	// noncetopo should be verified for all ring members simultaneously
 	// this can lead to tx rejection
 	// we currently bypass this since random members are chosen which have not been used in last 5 block
-	_, noncetopo, block_hash, self_e, err := w.GetEncryptedBalanceAtTopoHeight(zeroscid, -1, w.GetAddress().String())
+	_, noncetopo, _, self_e, err := w.GetEncryptedBalanceAtTopoHeight(zeroscid, -1, w.GetAddress().String())
 	if err != nil {
 		err = fmt.Errorf("could not obtain encrypted balance for self err %s\n", err)
 		return
@@ -236,7 +219,7 @@ func (w *Wallet_Memory) TransferPayload0(transfers []rpc.Transfer, ringsize uint
 		topoheight = daemon_topoheight - 3
 	}
 
-	_, _, block_hash, self_e, _ = w.GetEncryptedBalanceAtTopoHeight(transfers[0].SCID, topoheight, w.GetAddress().String())
+	_, _, _, self_e, _ = w.GetEncryptedBalanceAtTopoHeight(transfers[0].SCID, topoheight, w.GetAddress().String())
 	if err != nil {
 		fmt.Printf("self unregistered err %s\n", err)
 		return
